Add custom CSS and visual improvements
