+++ pages/3_📊_Data_Insights.py
@@ -4,9 +4,47 @@ import plotly.express as px
 import plotly.graph_objects as go
 from plotly.subplots import make_subplots
 import numpy as np
+from pathlib import Path
 
 # Page config
 st.set_page_config(page_title="Data Insights", page_icon="📊", layout="wide")
+
+# ============================================================================
+# LOAD CUSTOM CSS
+# ============================================================================
+
+def load_css():
+    """Load custom CSS styling."""
+    css_file = Path(__file__).parent.parent / "assets" / "style.css"
+    if css_file.exists():
+        with open(css_file) as f:
+            st.markdown(f"<style>{f.read()}</style>", unsafe_allow_html=True)
+
+load_css()
+
+# ============================================================================
+# SIDEBAR BRANDING
+# ============================================================================
+
+with st.sidebar:
+    st.markdown("""
+        <div class="logo-container">
+            <h1 style="color: white; font-size: 1.5rem; margin-bottom: 0;">🎓</h1>
+            <h2 style="color: white; font-size: 1.2rem; margin-top: 0;">Student Performance</h2>
+            <p style="color: rgba(255,255,255,0.8); font-size: 0.9rem;">AI-Powered Risk Predictor</p>
+        </div>
+    """, unsafe_allow_html=True)
+    
+    st.markdown("---")
+    st.markdown("### 👥 Team")
+    st.markdown("""
+    <small style="color: rgba(255,255,255,0.9);">
+    <b>Lead:</b> Pratik Mane<br>
+    <b>Members:</b> Emmanuel Atilola, Yugant Nagralawala, Hamza Almani<br>
+    <b>Advisor:</b> Prof. Leon Johnson
+    </small>
+    """, unsafe_allow_html=True)
+    st.markdown("---")
 
 # Title
 st.title("📊 Data Insights & Visualizations")
