--- pages/EDA.py
@@ -1,6 +1,0 @@
-import streamlit as st
-
-st.title("EDA")
-
-st.header("🚧 In progress. Come back later")
-
