--- pages/Prediction.py
@@ -1,5 +1,0 @@
-import streamlit as st
-
-st.title("Prediction")
-
-st.header("🚧 In progress. Come back later")
