--- pages/Visualization.py
@@ -1,5 +1,0 @@
-import streamlit as st
-
-st.title("Visualization")
-
-st.header("🚧 In progress. Come back later")
